# apps/api/api/openapi.yaml
components:
  schemas:
...
    handler.BudgetPeriodResponse:
      properties:
        budget:
          type: integer
        category_name:
          type: string
        created_at:
          type: string
        id:
          type: integer
        period:
          type: string
        rollover:
          type: integer
        spent:
          type: integer
        updated_at:
          type: string
      required:
      - budget
      - category_name
      - created_at
      - id
      - period
      - rollover
      - spent
      - updated_at
      type: object
    handler.BudgetWithProgressResponse:
      properties:
        category_name:
          type: string
        color:
          type: string
        monthly_budget:
          type: integer
        rollover:
          type: integer
        rollover_enabled:
          type: boolean
        spent:
          type: integer
      required:
      - category_name
      - color
      - monthly_budget
      - rollover
      - rollover_enabled
      - spent
      type: object
...
    handler.CreateGoalRequest:
      properties:
        color:
          type: string
        current_amount:
          type: integer
        deadline:
          type: string
        icon:
          type: string
        name:
          type: string
        target_amount:
          type: integer
      required:
      - name
      - target_amount
      type: object
...
    handler.DataWrapper:
      properties:
        data: {}
      type: object
...
    handler.InsightResponse:
      properties:
        action_url:
          type: string
        created_at:
          type: string
        description:
          type: string
        dismissed:
          type: boolean
        expires_at:
          type: string
        id:
          type: integer
        title:
          type: string
        type:
          type: string
      required:
      - created_at
      - description
      - dismissed
      - id
      - title
      - type
      type: object
...
    handler.SavingsGoalResponse:
      properties:
        archived:
          type: boolean
        color:
          type: string
        created_at:
          type: string
        current_amount:
          type: integer
        deadline:
          type: string
        icon:
          type: string
        id:
          type: integer
        name:
          type: string
        target_amount:
          type: integer
        updated_at:
          type: string
      required:
      - archived
      - color
      - created_at
      - current_amount
      - icon
      - id
      - name
      - target_amount
      - updated_at
      type: object
...
    handler.StatusResponse:
      properties:
        status:
          example: ok
          type: string
      required:
      - status
      type: object
...
    handler.UpdateGoalRequest:
      properties:
        archived:
          type: boolean
        color:
          type: string
        current_amount:
          type: integer
        deadline:
          type: string
        icon:
          type: string
        name:
          type: string
        target_amount:
          type: integer
      type: object
    handler.UpdateInstitutionRequest:
      properties:
        color:
          type: string
        display_name:
          type: string
        hidden:
          type: boolean
        icon:
          type: string
      type: object
...
    handler.UpsertBudgetRequest:
      properties:
        category_name:
          type: string
        color:
          type: string
        monthly_budget:
          type: integer
        rollover_enabled:
          type: boolean
      required:
      - category_name
      - monthly_budget
      type: object
info:
  contact: {}
  description: Personal finance API wrapping MoneyForward ME data
  title: Kanjo API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /api/v1/budgets:
    get:
      operationId: listBudgets
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/handler.DataWrapper'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/handler.BudgetWithProgressResponse'
                      type: array
                  type: object
          description: OK
      summary: List budgets
      tags:
      - budgets
    put:
      operationId: upsertBudget
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.UpsertBudgetRequest'
        description: Budget data
        required: true
        x-originalParamName: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.StatusResponse'
          description: OK
      summary: Create or update budget
      tags:
      - budgets
  /api/v1/budgets/{category}:
    delete:
      operationId: deleteBudget
      parameters:
      - description: Category name
        in: path
        name: category
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
      summary: Delete budget
      tags:
      - budgets
  /api/v1/budgets/{category}/periods:
    get:
      operationId: getBudgetPeriods
      parameters:
      - description: Category name
        in: path
        name: category
        required: true
        schema:
          type: string
      - description: Number of periods
        in: query
        name: limit
        schema:
          default: 12
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/handler.DataWrapper'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/handler.BudgetPeriodResponse'
                      type: array
                  type: object
          description: OK
      summary: Get budget periods
      tags:
      - budgets
...
  /api/v1/goals:
    get:
      operationId: listGoals
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/handler.DataWrapper'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/handler.SavingsGoalResponse'
                      type: array
                  type: object
          description: OK
      summary: List goals
      tags:
      - goals
    post:
      operationId: createGoal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.CreateGoalRequest'
        description: Goal data
        required: true
        x-originalParamName: body
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.SavingsGoalResponse'
          description: Created
      summary: Create goal
      tags:
      - goals
  /api/v1/goals/{id}:
    delete:
      operationId: deleteGoal
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      summary: Delete goal
      tags:
      - goals
    put:
      operationId: updateGoal
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.UpdateGoalRequest'
        description: Goal data
        required: true
        x-originalParamName: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.StatusResponse'
          description: OK
      summary: Update goal
      tags:
      - goals
  /api/v1/insights:
    get:
      operationId: listInsights
      parameters:
      - description: Number of results
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/handler.DataWrapper'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/handler.InsightResponse'
                      type: array
                  type: object
          description: OK
      summary: List insights
      tags:
      - insights
  /api/v1/insights/{id}/dismiss:
    post:
      operationId: dismissInsight
      parameters:
      - description: Insight ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.StatusResponse'
          description: OK
      summary: Dismiss insight
      tags:
      - insights
...
  /api/v1/institutions/{name}:
    patch:
      operationId: updateInstitution
      parameters:
      - description: Institution name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.UpdateInstitutionRequest'
        description: Institution data
        required: true
        x-originalParamName: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.StatusResponse'
          description: OK
      summary: Update institution
      tags:
      - institutions
...
servers:
- url: http://localhost:8080/